Realizada mudanças com divisoes de classes e novas funcoes
