+++ pyproject.toml
@@ -7,6 +7,7 @@ authors = ["Your Name <you@example.com>"]
 [tool.poetry.dependencies]
 python = "^3.8"
 discord = "^1.0.1"
+flask = "^1.1.2"
 
 [tool.poetry.dev-dependencies]
